Update HttpConnection.cs
The check Type.GetType("Mono.Runtime") == null evaluates to true on the Xamarin platform, but HttpWebRequest.DefaultMaximumErrorResponseLength will throw a NotImplementedException. Xamarin is built on Mono, but the following thread would indicate that there may be some issues with the Mono.Runtime type being linked out: http://forums.xamarin.com/discussion/1551/detect-mono-using-mono-runtime-class.

The proposed solution to make a check that works on Xamarin as well is to include "System.MonoType" in the evaluation, which apparently will not be linked out.

I tested this on iOS only. It is possible that when running on Android devices the current check will work. Also, this issue only occurs when running on a physical iOS device. When running in an iOS simulator, apparently the Mono.Runtime is not linked out, so the current check works.